Added counters to all applications, added adapter trimmer (same logic as overlapper, but will return PE reads), and added stats program

+++ common/src/typedefs.h
@@ -1,7 +1,7 @@
 #ifndef TYPDEF_H
 #define TYPDEF_H
 
-typedef std::unordered_map <std::string, size_t> Counter;
+typedef std::unordered_map <std::string, unsigned long long int> Counter;
 class Counters;
 
 #endif
